% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babynamer-package.R
\docType{package}
\name{babynamer-package}
\alias{babynamer}
\alias{babynamer-package}
\title{babynamer}
\description{
Tools to access baby name data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abresler/babynamer}
  \item Report bugs at \url{https://github.com/abresler/babynamer/issues}
}

}
\author{
Alex Bresler (abresler@asbcllc.com)
}
\keyword{internal}
