% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{us_baby_names}
\alias{us_baby_names}
\title{United States Baby Name data}
\usage{
us_baby_names(type = "national", include_features = T)
}
\arguments{
\item{type}{the type of data options include \itemize{
\item `national` - National popularity by state, sex and year - default
\item `state` - Popularity by state, sex and year
\item `territory` -  Popularity by territory, sex and year
}}

\item{include_features}{if `TRUE` includes features about the popularity of the name by sex}
}
\value{
a `tibble` or a nested `tibble` if `type` length exceeds 1
}
\description{
Returns detailed data about the popularity of
various baby names in the United States.
}
\details{
Includes access to national, state, and territorial
data.
}
\examples{
\dontrun{
library(babynamer)

tbl_usa <- us_baby_names(type = "national")
tbl_state <- us_baby_names(type = "state")
tbl_territory <- us_baby_names(type = "territory")
us_baby_names <- us_baby_names(type = c("national","state", "territory"))

}
}
